Introduce TMCReplay and restructure repo (#22)
* Introduce TMCReplay and restructure repo

[TMCReplay]

Simulation engine based on the Virtual Monte Carlo (VMC) framework to
exactly replay the steps as previously recorded from another run using
the MCStepLogger.

Cut and process settings can be varied to study the impact on the steps.
That funcionality is currently not fully implemented but will be as the
next step.

The replay engine is built by default but the build can be switched off
setting `-MCStepLogger_BUILD_TMCReplay=OFF" at configure time when
`CMake` is invoked.

In the easiest case it can be run using the installed executable
`tmcreplay`. Use `tmcreplay --help` to see all options.

Generally, the engine can be used as usual within any other simulation
framework. In that case it might be used to evaluate the impact of
different cut and process settings on hist or digits.

[MCStepLogger]

3-momenta (px, py, pz) added to logging as that is necessary to replay
simulation correctly

[Repository restructuring]

The original MCStepLogger (and the TMCReplay engine) were moved to
corresponding sub-directories.

Installation of libraries was unified through common CMake function.

[Apply clang-format to cxx (previously not done)]

* switch off test builds by default

* Complete TVirtualMC interface implementations

* implement

  * TVirtualMC::Gs* methods

  * TVirtualMC::{Material,Mixture,Medium}

-> overall implementation follows
   https://github.com/vmc-project/vmc/blob/master/source/include/TGeoMCGeometry.h
   however, in this case with a cached pointer to TGeoManager instead of
   everytime obtaining it from gGeoManager

* small updates of READMEs and executable source replay.cxx

* update tracking and stepping logic

* ignore additional secondaries secondaries pushed during hit creation
  (or anywhere during TVirtualMCApplication::Stepping)

* take copy number into account when updating the TGeoManager/TNavigator
  state

* Fix issues with running Replay stand-alone

* add MCReplayGenericApplication to be interceptable

* update replay test

* Add energy cut option to replay executable

Co-authored-by: Benedikt Volkel <benedikt.volkel@cern.ch>